requirements.txt
opencv-python==4.7.0.72
Pillow==11.3.0
numpy==1.23.5
pytesseract==0.3.13
openai==1.109.0
python-dotenv==1.0.1
tensorflow==2.12.0